Move constants to separate file to avoid circular imports
