+++ src/ros_vision_interaction/scripts/ros_interaction_manager.py
@@ -7,10 +7,10 @@ import pymongo
 import rospy
 
 from controllers import InteractionManager
-from controllers.vision_project_delegator import DatabaseKeys, INITIAL_STATE_DB
 from interaction_builder import InteractionBuilder
 from interfaces import CordialInterface
 from vision_project_tools import init_db
+from vision_project_tools.constants import DatabaseKeys, INITIAL_STATE_DB
 from vision_project_tools.engine_statedb import EngineStateDb as StateDb
 
 from ros_vision_interaction.msg import StartInteractionAction, StartInteractionResult
